+++ zinnia/plugins/models.py
@@ -41,6 +41,12 @@ class LatestEntriesPlugin(CMSPlugin):
         the template_to_render attribute"""
         return self.template_to_render
 
+    def copy_relations(self, old_instance):
+        """Duplicate ManyToMany relations on plugin copy"""
+        self.tags = old_instance.tags.all()
+        self.authors = old_instance.authors.all()
+        self.categories = old_instance.categories.all()
+
     def __unicode__(self):
         return _('%s entries') % self.number_of_entries
 
@@ -59,8 +65,24 @@ class SelectedEntriesPlugin(CMSPlugin):
         the template_to_render attribute"""
         return self.template_to_render
 
+    def copy_relations(self, old_instance):
+        """Duplicate ManyToMany relations on plugin copy"""
+        self.entries = old_instance.entries.all()
+
     def __unicode__(self):
         return _('%s entries') % self.entries.count()
+
+
+class RandomEntriesPlugin(CMSPlugin):
+    """CMS Plugin for displaying random entries"""
+
+    number_of_entries = models.IntegerField(_('number of entries'), default=5)
+    template_to_render = models.CharField(_('template'), blank=True,
+                                          max_length=250, choices=TEMPLATES,
+                                          help_text=_('Template used to display the plugin'))
+
+    def __unicode__(self):
+        return _('%s entries') % self.number_of_entries
 
 
 def invalidate_menu_cache(sender, **kwargs):