+++ docs/CHANGELOG.rst
@@ -8,7 +8,9 @@ CHANGELOG
 * Using Microformats
 * Mails for comment reply
 * Entry model can be extended
+* More plugins for django-cms
 * Zinnia to Wordpress utility command
+* Code cleaning and optimizations
 * And a lot of bug fixes
 
 0.7
@@ -20,7 +22,7 @@ CHANGELOG
 * Private posts
 * Hierarchical categories
 * TinyMCE integration
-* Code optimisation
+* Code optimizations
 * And a lot of bug fixes
 
 0.6